refactor: Domain 패키지에서 VO 객체에 대해 패키지 분리
